# api-inventory/src/main/java/com/demo/api_inventory/dto/product/Product.java
package com.demo.api_inventory.dto.product;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.time.LocalDate;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class Product {

    private Long id;

    private String name;

    private float price;

    private int antique;

}